[macOS] Move HID console output to separate window
